in progress - Variable Manager overhaul
