+++ Assets/Scripts/Time/VariableLibrary.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 90e2e678da0836f438503744f50d8bae
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
