+++ Assets/Scripts/Time/VariableLibrary/VariableLibrary.cs
@@ -1,0 +1,25 @@
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Time
+{
+    [CreateAssetMenu(fileName = "VariableLibrary", menuName = "Scriptable Objects/VariableLibrary")]
+    public class VariableLibrary : ScriptableObject
+    {
+        [SerializeField]
+        private VarMapping[] mappings;
+
+        // Added method to expose the mappings for initialization
+        public VarMapping[] GetMappings()
+        {
+            return mappings;
+        }
+
+        [System.Serializable]
+        public class VarMapping // Made public to be accessible from RuntimeVariableLibrary
+        {
+            public string variableName;
+            public float initialValue;
+        }
+    }
+}
